requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
firebase-admin==6.2.0
requests==2.31.0
python-dateutil==2.8.2